update compile config

+++ src/game/block_selection.c
@@ -54,7 +54,7 @@ void block_selection_draw(us_t now) {
     GX_SetTevOrder(GX_TEVSTAGE0, GX_TEXCOORDNULL, GX_TEXMAP_NULL, GX_COLOR0A0);
     GX_SetTevOp(GX_TEVSTAGE0, GX_PASSCLR);
 
-    u8 alpha = 0x5f + (sinf(now / 150000.0f) * 0x10);
+    u8 alpha = 0x5f + (u8) (sinf((f32) now / 150000.0f) * 0x10);
     GX_SetTevColor(GX_TEVREG1, (GXColor){ 0xff, 0xff, 0xff, alpha });
     GX_SetTevColorIn(GX_TEVSTAGE0, GX_CC_ZERO, GX_CC_ZERO, GX_CC_ZERO, GX_CC_C1);
     GX_SetTevColorOp(GX_TEVSTAGE0, GX_TEV_ADD, GX_TB_ZERO, GX_CS_SCALE_1, GX_TRUE, GX_TEVPREV);
@@ -100,7 +100,7 @@ void block_selection_handle_location(Mtx view, world_location_t location) {
     u8vec3s block_pos = location.bl_pos;
 
     guMtxIdentity(model);
-    guMtxTransApply(model, model, chunk_pos.x * NUM_ROW_BLOCKS_PER_BLOCK_CHUNK, chunk_pos.y * NUM_ROW_BLOCKS_PER_BLOCK_CHUNK, chunk_pos.z * NUM_ROW_BLOCKS_PER_BLOCK_CHUNK);
+    guMtxTransApply(model, model, (f32) chunk_pos.x * NUM_ROW_BLOCKS_PER_BLOCK_CHUNK, (f32) chunk_pos.y * NUM_ROW_BLOCKS_PER_BLOCK_CHUNK, (f32) chunk_pos.z * NUM_ROW_BLOCKS_PER_BLOCK_CHUNK);
     guMtxConcat(view, model, model_view);
     
     GX_LoadPosMtxImm(model_view, MATRIX_INDEX);
